requirements.txt
streamlit
pytesseract
pdf2image
python-dateutil
pillow
requests